requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
cffi==1.15.1
charset-normalizer==3.3.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==3.2.22
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
frozenlist==1.3.3
h11==0.14.0
httpie==3.2.2
idna==3.4
importlib-metadata==6.7.0
kombu==5.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==1.3.14
numpy==1.21.6
openai==0.28.1
outcome==1.2.0
pandas==1.3.5
prompt-toolkit==3.0.39
protobuf==3.20.3
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
wincertstore==0.2
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
